# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nestly AI Home Search API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyGeocodeStatus = typeof PropertyGeocodeStatus[keyof typeof PropertyGeocodeStatus];


export const PropertyGeocodeStatus = {
  ok: 'ok',
  pending: 'pending',
  failed: 'failed',
} as const;

export interface Property {
  id: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
  price: number;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  sqft?: number | null;
  propertyType: string;
  /** @nullable */
  yearBuilt?: number | null;
  /** @nullable */
  description?: string | null;
  photos: string[];
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  geocodeStatus?: PropertyGeocodeStatus;
  /** @nullable */
  mlsId?: string | null;
  /** @nullable */
  listingUrl?: string | null;
  /** @nullable */
  daysOnMarket?: number | null;
  /** @nullable */
  pricePerSqft?: number | null;
  source: string;
  /** @nullable */
  matchScore?: number | null;
}

export interface GeocodeStatusInput {
  ids: string[];
}

export type GeocodeStatusEntryGeocodeStatus = typeof GeocodeStatusEntryGeocodeStatus[keyof typeof GeocodeStatusEntryGeocodeStatus];


export const GeocodeStatusEntryGeocodeStatus = {
  ok: 'ok',
  pending: 'pending',
  failed: 'failed',
} as const;

export interface GeocodeStatusEntry {
  id: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  geocodeStatus: GeocodeStatusEntryGeocodeStatus;
}

export interface PropertySearchInput {
  prompt: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  maxResults?: number | null;
}

export interface ParsedFilters {
  /** @nullable */
  location?: string | null;
  /** @nullable */
  minPrice?: number | null;
  /** @nullable */
  maxPrice?: number | null;
  /** @nullable */
  minBedrooms?: number | null;
  /** @nullable */
  minBathrooms?: number | null;
  /** @nullable */
  propertyType?: string | null;
  /** @nullable */
  minSqft?: number | null;
  /** @nullable */
  maxSqft?: number | null;
  keywords?: string[];
}

export interface PropertySearchResult {
  properties: Property[];
  totalCount: number;
  searchSummary: string;
  parsedFilters: ParsedFilters;
}

export interface FavoriteProperty {
  id: number;
  propertyData: Property;
  createdAt: string;
}

export interface FavoriteInput {
  propertyData: Property;
}

export interface TrendingSearch {
  query: string;
  count: number;
  /** @nullable */
  location?: string | null;
}

export interface MarketInsights {
  location: string;
  averagePrice: number;
  medianPrice: number;
  activeListings: number;
  avgDaysOnMarket: number;
  pricePerSqft: number;
  priceChangePercent: number;
  hotNeighborhoods: string[];
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export interface SavedSearch {
  id: number;
  /** @nullable */
  name?: string | null;
  query: string;
  parsedFilters?: ParsedFilters;
  alertEnabled: boolean;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  lastResultCount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface SavedSearchInput {
  query: string;
  /** @nullable */
  name?: string | null;
  parsedFilters?: ParsedFilters;
  alertEnabled?: boolean;
}

export interface VisitedProperty {
  id: number;
  propertyData: Property;
  visitCount: number;
  firstVisitedAt: string;
  lastVisitedAt: string;
}

export interface VisitedPropertyInput {
  propertyData: Property;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  href?: string | null;
  /** @nullable */
  readAt?: string | null;
  createdAt: string;
}

export interface UnreadCount {
  count: number;
}

export type GetMarketInsightsParams = {
location?: string;
};