# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nestly AI Home Search API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/search:
    post:
      operationId: searchProperties
      tags: [properties]
      summary: AI-powered property search from natural language prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertySearchInput"
      responses:
        "200":
          description: Search results with matched properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertySearchResult"
...
  /properties/geocode-status:
    post:
      operationId: getGeocodeStatus
      tags: [properties]
      summary: Resolve geocode coordinates/status for property IDs (polled by the map)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeocodeStatusInput"
      responses:
        "200":
          description: Geocode status for the requested properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeocodeStatusEntry"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List all saved favorite properties
      responses:
        "200":
          description: List of favorites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoriteProperty"
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Save a property to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Created favorite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteProperty"
  /favorites/{id}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a property from favorites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /saved-searches:
    get:
      operationId: listSavedSearches
      tags: [saved-searches]
      summary: List the authenticated user's saved searches
      responses:
        "200":
          description: List of saved searches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedSearch"
    post:
      operationId: createSavedSearch
      tags: [saved-searches]
      summary: Save a search for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedSearchInput"
      responses:
        "201":
          description: Created saved search
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
  /saved-searches/{id}:
    delete:
      operationId: deleteSavedSearch
      tags: [saved-searches]
      summary: Delete a saved search
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /saved-searches/{id}/run:
    post:
      operationId: runSavedSearch
      tags: [saved-searches]
      summary: Re-run a saved search and return fresh real listings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fresh search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertySearchResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Marked read
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        price:
          type: number
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["number", "null"]
        sqft:
          type: ["integer", "null"]
        propertyType:
          type: string
        yearBuilt:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        photos:
          type: array
          items:
            type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        geocodeStatus:
          type: string
          enum: [ok, pending, failed]
        mlsId:
          type: ["string", "null"]
        listingUrl:
          type: ["string", "null"]
        daysOnMarket:
          type: ["integer", "null"]
        pricePerSqft:
          type: ["number", "null"]
        source:
          type: string
        matchScore:
          type: ["number", "null"]
      required:
        - id
        - address
        - city
        - state
        - zipCode
        - price
        - propertyType
        - photos
        - source
    GeocodeStatusInput:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
      required:
        - ids
    GeocodeStatusEntry:
      type: object
      properties:
        id:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        geocodeStatus:
          type: string
          enum: [ok, pending, failed]
      required:
        - id
        - geocodeStatus
    PropertySearchInput:
      type: object
      properties:
        prompt:
          type: string
        location:
          type: ["string", "null"]
        maxResults:
          type: ["integer", "null"]
      required:
        - prompt
    ParsedFilters:
      type: object
      properties:
        location:
          type: ["string", "null"]
        minPrice:
          type: ["number", "null"]
        maxPrice:
          type: ["number", "null"]
        minBedrooms:
          type: ["integer", "null"]
        minBathrooms:
          type: ["number", "null"]
        propertyType:
          type: ["string", "null"]
        minSqft:
          type: ["integer", "null"]
        maxSqft:
          type: ["integer", "null"]
        keywords:
          type: array
          items:
            type: string
    PropertySearchResult:
      type: object
      properties:
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        totalCount:
          type: integer
        searchSummary:
          type: string
        parsedFilters:
          $ref: "#/components/schemas/ParsedFilters"
      required:
        - properties
        - totalCount
        - searchSummary
        - parsedFilters
    FavoriteProperty:
      type: object
      properties:
        id:
          type: integer
        propertyData:
          $ref: "#/components/schemas/Property"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - propertyData
        - createdAt
    FavoriteInput:
      type: object
      properties:
        propertyData:
          $ref: "#/components/schemas/Property"
      required:
        - propertyData
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SavedSearch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: ["string", "null"]
        query:
          type: string
        parsedFilters:
          $ref: "#/components/schemas/ParsedFilters"
        alertEnabled:
          type: boolean
        lastRunAt:
          type: ["string", "null"]
          format: date-time
        lastResultCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - query
        - alertEnabled
        - createdAt
        - updatedAt
    SavedSearchInput:
      type: object
      properties:
        query:
          type: string
        name:
          type: ["string", "null"]
        parsedFilters:
          $ref: "#/components/schemas/ParsedFilters"
        alertEnabled:
          type: boolean
      required:
        - query
...