# lib/api-zod/src/generated/types/geocodeStatusEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nestly AI Home Search API
 * OpenAPI spec version: 0.1.0
 */
import type { GeocodeStatusEntryGeocodeStatus } from './geocodeStatusEntryGeocodeStatus';

export interface GeocodeStatusEntry {
  id: string;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  geocodeStatus: GeocodeStatusEntryGeocodeStatus;
}