# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nestly AI Home Search API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyGeocodeStatus } from './propertyGeocodeStatus';

export interface Property {
  id: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
  price: number;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  sqft?: number | null;
  propertyType: string;
  /** @nullable */
  yearBuilt?: number | null;
  /** @nullable */
  description?: string | null;
  photos: string[];
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  geocodeStatus?: PropertyGeocodeStatus;
  /** @nullable */
  mlsId?: string | null;
  /** @nullable */
  listingUrl?: string | null;
  /** @nullable */
  daysOnMarket?: number | null;
  /** @nullable */
  pricePerSqft?: number | null;
  source: string;
  /** @nullable */
  matchScore?: number | null;
}